Change Dwarf_CU_Context files list from vector to map
The current dwarf parser for OSX does not actually prune duplicate file names,
which leads to wasted time parsing the file name list later on. Switching the
container for the file names from list to map would allow us to remove the
duplicate file names from the list and thus improve the performance of ddrgen.
On AIX, we keep a separate set for this purpose and this change would also do
away with that. We need a map instead of a set so that dwarf_srcfiles returns
the file list in the correct order indicated by the DW_AT_decl_file attribute.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>